refactor: move federated server logic to its own service
