+++ core/p2p/p2p_disabled.go
@@ -14,6 +14,10 @@ func GenerateToken() string {
 	return "not implemented"
 }
 
+func (f *FederatedServer) Start(ctx context.Context) error {
+	return fmt.Errorf("not implemented")
+}
+
 func ServiceDiscoverer(ctx context.Context, node *node.Node, token, servicesID string, fn func(string, NodeData)) error {
 	return fmt.Errorf("not implemented")
 }
